refactor(MultipleSelectionFloatingPanelViewController): Reduce complexity, preserving existing logic
