+++ kDrive/UI/Controller/Files/MultipleSelectionFloatingPanelViewController.swift
@@ -23,7 +23,7 @@ import kDriveCore
 import kDriveResources
 import UIKit
 
-class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
+final class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
     @LazyInjectService var accountManager: AccountManageable
 
     var driveFileManager: DriveFileManager!
@@ -35,6 +35,12 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
     var downloadInProgress = false
     var reloadAction: (() -> Void)?
 
+    var downloadedArchiveUrl: URL?
+    var success = true
+    var addAction = true
+    var currentArchiveId: String?
+    var downloadError: DriveError?
+
     weak var presentingParent: UIViewController?
 
     var sharedWithMe: Bool {
@@ -42,12 +48,6 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
     }
 
     var actions = FloatingPanelAction.listActions
-
-    private var addAction = true
-    private var success = true
-    private var downloadedArchiveUrl: URL?
-    private var currentArchiveId: String?
-    private var downloadError: DriveError?
 
     convenience init() {
         self.init(collectionViewLayout: MultipleSelectionFloatingPanelViewController.createLayout())
@@ -76,220 +76,10 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
         }
     }
 
-    func handleAction(_ action: FloatingPanelAction, at indexPath: IndexPath) {
-        let action = actions[indexPath.row]
-        success = true
-        addAction = true
-        let group = DispatchGroup()
-
-        switch action {
-        case .offline:
-            addAction = FileActionsHelper.offline(files: files, driveFileManager: driveFileManager, group: group) {
-                self.downloadInProgress = true
-                self.collectionView.reloadItems(at: [indexPath])
-            } completion: { file, error in
-                if error != nil {
-                    self.success = false
-                }
-                if let file = self.driveFileManager.getCachedFile(id: file.id) {
-                    self.changedFiles?.append(file)
-                }
-            }
-        case .favorite:
-            group.enter()
-            Task {
-                let isFavored = try await FileActionsHelper.favorite(files: files,
-                                                                     driveFileManager: driveFileManager,
-                                                                     completion: favorite)
-                addAction = isFavored
-                group.leave()
-            }
-        case .manageCategories:
-            FileActionsHelper.manageCategories(files: files,
-                                               driveFileManager: driveFileManager,
-                                               from: self,
-                                               group: group,
-                                               presentingParent: presentingParent)
-        case .folderColor:
-            FileActionsHelper.folderColor(files: files,
-                                          driveFileManager: driveFileManager,
-                                          from: self,
-                                          presentingParent: presentingParent,
-                                          group: group) { isSuccess in
-                self.success = isSuccess
-            }
-        case .download:
-            if !allItemsSelected &&
-                (files.allSatisfy { $0.convertedType == .image || $0.convertedType == .video } || files.count <= 1) {
-                for file in files {
-                    if file.isDownloaded {
-                        FileActionsHelper.save(file: file, from: self, showSuccessSnackBar: false)
-                    } else {
-                        guard let observerViewController = view.window?.rootViewController else { return }
-                        downloadInProgress = true
-                        collectionView.reloadItems(at: [indexPath])
-                        group.enter()
-                        DownloadQueue.instance
-                            .observeFileDownloaded(observerViewController, fileId: file.id) { [weak self] _, error in
-                                guard let self else { return }
-                                if error == nil {
-                                    Task { @MainActor in
-                                        FileActionsHelper.save(file: file, from: self, showSuccessSnackBar: false)
-                                    }
-                                } else {
-                                    success = false
-                                }
-                                group.leave()
-                            }
-                        DownloadQueue.instance.addToQueue(file: file, userId: accountManager.currentUserId)
-                    }
-                }
-            } else {
-                if downloadInProgress,
-                   let currentArchiveId,
-                   let operation = DownloadQueue.instance.archiveOperationsInQueue[currentArchiveId] {
-                    group.enter()
-                    let alert = AlertTextViewController(
-                        title: KDriveResourcesStrings.Localizable.cancelDownloadTitle,
-                        message: KDriveResourcesStrings.Localizable.cancelDownloadDescription,
-                        action: KDriveResourcesStrings.Localizable.buttonYes,
-                        destructive: true
-                    ) {
-                        operation.cancel()
-                        self.downloadError = .taskCancelled
-                        self.success = false
-                        group.leave()
-                    }
-                    present(alert, animated: true)
-                } else {
-                    downloadedArchiveUrl = nil
-                    downloadInProgress = true
-                    collectionView.reloadItems(at: [indexPath])
-                    group.enter()
-                    downloadArchivedFiles(downloadCellPath: indexPath) { result in
-                        switch result {
-                        case .success(let archiveUrl):
-                            self.downloadedArchiveUrl = archiveUrl
-                            self.success = true
-                        case .failure(let error):
-                            self.downloadError = error
-                            self.success = false
-                        }
-                        group.leave()
-                    }
-                }
-            }
-        case .move:
-            FileActionsHelper.move(files: files,
-                                   exceptFileIds: exceptFileIds ?? [],
-                                   from: currentDirectory,
-                                   allItemsSelected: allItemsSelected,
-                                   observer: self,
-                                   driveFileManager: driveFileManager) { [weak self] viewController in
-                dismiss(animated: true) {
-                    self?.presentingParent?.present(viewController, animated: true)
-                }
-            }
-        case .duplicate:
-            let selectFolderNavigationController = SelectFolderViewController.instantiateInNavigationController(
-                driveFileManager: driveFileManager,
-                disabledDirectoriesSelection: files.compactMap(\.parent)
-            ) { [files = files.map { $0.freezeIfNeeded() }] selectedDirectory in
-                Task {
-                    do {
-                        try await self.copy(files: files, to: selectedDirectory)
-                    } catch {
-                        self.success = false
-                    }
-                }
-                group.leave()
-            }
-            group.enter()
-            present(selectFolderNavigationController, animated: true)
-            collectionView.reloadItems(at: [indexPath])
-        default:
-            break
-        }
-
-        group.notify(queue: .main) {
-            if self.success {
-                switch action {
-                case .offline:
-                    let filesUpdatedNumber = self.files.filter { !$0.isDirectory }.count
-                    if self.addAction {
-                        UIConstants
-                            .showSnackBar(message: KDriveResourcesStrings.Localizable
-                                .fileListAddOfflineConfirmationSnackbar(filesUpdatedNumber))
-                    } else {
-                        UIConstants
-                            .showSnackBar(message: KDriveResourcesStrings.Localizable
-                                .fileListRemoveOfflineConfirmationSnackbar(filesUpdatedNumber))
-                    }
-                case .favorite:
-                    if self.addAction {
-                        UIConstants
-                            .showSnackBar(message: KDriveResourcesStrings.Localizable
-                                .fileListAddFavoritesConfirmationSnackbar(self.files.count))
-                    } else {
-                        UIConstants
-                            .showSnackBar(message: KDriveResourcesStrings.Localizable
-                                .fileListRemoveFavoritesConfirmationSnackbar(self.files.count))
-                    }
-                case .folderColor:
-                    UIConstants
-                        .showSnackBar(message: KDriveResourcesStrings.Localizable
-                            .fileListColorFolderConfirmationSnackbar(self.files.filter(\.canBeColored).count))
-                case .duplicate:
-                    guard self.addAction else { break }
-                    UIConstants
-                        .showSnackBar(message: KDriveResourcesStrings.Localizable
-                            .fileListDuplicationConfirmationSnackbar(self.files.count))
-                case .download:
-                    guard !self.files.isEmpty && self.files
-                        .allSatisfy({ $0.convertedType == .image || $0.convertedType == .video }) else { break }
-                    if self.files.count <= 1, let file = self.files.first {
-                        let message = file.convertedType == .image
-                            ? KDriveResourcesStrings.Localizable.snackbarImageSavedConfirmation
-                            : KDriveResourcesStrings.Localizable.snackbarVideoSavedConfirmation
-                        UIConstants.showSnackBar(message: message)
-                    } else {
-                        UIConstants
-                            .showSnackBar(message: KDriveResourcesStrings.Localizable.snackBarImageVideoSaved(self.files.count))
-                    }
-                default:
-                    break
-                }
-            } else {
-                UIConstants.showSnackBarIfNeeded(error: self.downloadError ?? DriveError.unknownError)
-            }
-            self.files = self.changedFiles ?? []
-            self.downloadInProgress = false
-            self.collectionView.reloadItems(at: [indexPath])
-            if action == .download {
-                if let downloadedArchiveUrl = self.downloadedArchiveUrl {
-                    // Present from root view controller if the panel is no longer presented
-                    let viewController = self.view.window != nil
-                        ? self
-                        : (UIApplication.shared.delegate as! AppDelegate).topMostViewController
-                    guard viewController as? UIDocumentPickerViewController == nil else { return }
-                    let documentExportViewController = UIDocumentPickerViewController(
-                        url: downloadedArchiveUrl,
-                        in: .exportToService
-                    )
-                    viewController?.present(documentExportViewController, animated: true)
-                }
-            } else {
-                self.dismiss(animated: true)
-            }
-            self.reloadAction?()
-            self.changedFiles = []
-        }
-    }
-
     // MARK: - Private methods
 
     @MainActor
-    private func copy(files: [File], to selectedDirectory: File) async throws {
+    func copy(files: [File], to selectedDirectory: File) async throws {
         if files.count > Constants.bulkActionThreshold || allItemsSelected {
             // addAction = false // Prevents the snackbar to be displayed
             let action: BulkAction
@@ -322,7 +112,7 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
         }
     }
 
-    private func downloadArchivedFiles(downloadCellPath: IndexPath, completion: @escaping (Result<URL, DriveError>) -> Void) {
+    func downloadArchivedFiles(downloadCellPath: IndexPath, completion: @escaping (Result<URL, DriveError>) -> Void) {
         Task { [proxyFiles = files.map { $0.proxify() }, currentProxyDirectory = currentDirectory.proxify()] in
             do {
                 let archiveBody: ArchiveBody
@@ -364,7 +154,7 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
         }
     }
 
-    private func favorite(file: File) async {
+    func favorite(file: File) async {
         if let file = driveFileManager.getCachedFile(id: file.id) {
             await MainActor.run {
                 self.changedFiles?.append(file)
